Move endpoints to the separate files
