+++ ph/main.py
@@ -2,10 +2,13 @@ from fastapi import FastAPI, Request, status
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
-from starlette.templating import _TemplateResponse
+from starlette.templating import _TemplateResponse  # noqa
 
 from ph.models.info_response import VersionResponse
-from ph.routers.api import api
+from ph.routers.api.create import api as api_create
+from ph.routers.api.hex import api as api_hex
+from ph.routers.api.info import api as api_info
+from ph.routers.api.packets import api as api_packets
 
 app = FastAPI(
     title="Packet Helper Next",
@@ -24,7 +27,10 @@ app.add_middleware(
     allow_headers=["*"],
 )
 app.mount("/static", StaticFiles(directory="static"), name="static")
-app.include_router(api, prefix="/api", tags=["api"])
+
+router_api_config = {"prefix": "/api", "tags": ["api"]}
+for api_router in (api_create, api_hex, api_info, api_packets):
+    app.include_router(api_router, **router_api_config)
 
 templates = Jinja2Templates(directory="static")
